requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
build==1.4.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cleo==2.1.0
crashtest==0.4.1
distlib==0.4.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
dulwich==1.1.0
fastjsonschema==2.21.2
filelock==3.25.2
findpython==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.1.1
jaraco.functools==4.4.0
keyring==25.7.0
librt==0.8.1
more-itertools==10.8.0
msgpack==1.1.2
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pbs-installer==2026.3.10
pkginfo==1.12.1.2
platformdirs==4.9.4
poetry==2.3.2
poetry-core==2.3.1
pycparser==3.0
pydantic_core==2.41.5
pyproject_hooks==1.2.0
python-discovery==1.1.3
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
shellingham==1.5.4
sqlparse==0.5.5
tomlkit==0.14.0
trove-classifiers==2026.1.14.14
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==21.2.0
xattr==1.3.0
zstandard==0.25.0
